% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{filter_split_sce}
\alias{filter_split_sce}
\title{intake labelled SCE and filter using dropletUtils}
\usage{
filter_split_sce(sce_split, output_folder, exp_name)
}
\value{
a SCE filtered on inctact cell barcodes
}
\description{
intake labelled SCE and filter using dropletUtils
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
