% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{gen_split_sce}
\alias{gen_split_sce}
\title{generate the sce}
\usage{
gen_split_sce(
  sub_lib_fp,
  output_folder,
  inex_mode = "inex",
  ds_dge_mtx = "all",
  dge_mtx,
  gene_names,
  exp_name,
  library_stats_df
)
}
\value{
SCE output and reports from the split-seq run
}
\description{
generate the sce
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
