% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{generate_barcoding_heatmaps}
\alias{generate_barcoding_heatmaps}
\title{}
\usage{
generate_barcoding_heatmaps(sce_split, output_folder, exp_name)
}
\value{
Null - generates the barcoding plate distribution heatmaps
}
\description{

}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
