% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{get_read_count}
\alias{get_read_count}
\title{extract total raw reads}
\usage{
get_read_count(fastq_path)
}
\arguments{
\item{fastq_path}{A string representing the raw FastQ path}
}
\value{
Total reads from raw fastq file
}
\description{
extract total raw reads
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
