% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{get_seq_run_info}
\alias{get_seq_run_info}
\title{extract total raw reads}
\usage{
get_seq_run_info(total_reads, sub_lib_fp, output_folder, exp_name)
}
\arguments{
\item{fastq_path}{A string representing the raw FastQ path}
}
\value{
Total reads from raw fastq file
}
\description{
extract total raw reads
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
