% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{label_sce_cdata}
\alias{label_sce_cdata}
\title{intake raw SCE and label with cell metadata as colData}
\usage{
label_sce_cdata(
  sce_split,
  rt_bc_map,
  lig_bc,
  sample_map,
  output_folder,
  exp_name
)
}
\value{
a SCE with
}
\description{
intake raw SCE and label with cell metadata as colData
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
