% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{label_sce_data}
\alias{label_sce_data}
\title{intake raw SCE and label with cell metadata as colData}
\usage{
label_sce_data(
  sce_split,
  rt_bc_map,
  lig_bc,
  sample_map,
  output_folder,
  exp_name,
  sl_index
)
}
\arguments{
\item{sce_split}{a SCE object representing the processed sublibrary}
}
\value{
a SCE with labelled barcode well and sample info integrated.
}
\description{
intake raw SCE and label with cell metadata as colData
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
