% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_merge_utils.R
\name{merge_sce_sublibs}
\alias{merge_sce_sublibs}
\title{Merge the individually processed unfiltered sublibraries}
\usage{
merge_sce_sublibs(sl_list)
}
\arguments{
\item{sl_list}{a list object containing the sublibraries to be merged}
}
\value{
SCE output of the merged data object
}
\description{
Merge the individually processed unfiltered sublibraries
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
