% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_merge_utils.R
\name{merge_sparse}
\alias{merge_sparse}
\title{Merge sparse matrices for SCE merging}
\usage{
merge_sparse(mtx_list)
}
\arguments{
\item{sce_list}{a list of sparse matrices to merge into one}
}
\value{
a combined sparse matrix object
}
\description{
Merge sparse matrices for SCE merging
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
