% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_merge_utils.R
\name{sce_merge_stats}
\alias{sce_merge_stats}
\title{create sequencing run statistics for merged sublibraries}
\usage{
sce_merge_stats(sce_split, output_folder, exp_name)
}
\arguments{
\item{sce_split}{An input merged SCE experiment object}
}
\value{
SCE object with seq statistics embedded in the SCE metadata
}
\description{
create sequencing run statistics for merged sublibraries
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
