% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{split_stats_output}
\alias{split_stats_output}
\title{create sequencing run statistics for each sublibrary}
\usage{
split_stats_output(sce_split, output_folder, exp_name)
}
\arguments{
\item{sce_split}{An input SCE experiment object}
}
\value{
SCE object with seq statistics embedded in the SCE metadata
}
\description{
create sequencing run statistics for each sublibrary
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
