% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{split_stats_downsample_rds}
\alias{split_stats_downsample_rds}
\title{generate the sce downsampling statistics}
\usage{
split_stats_downsample_rds(sce_split, output_folder, exp_name)
}
\arguments{
\item{sce_split}{a SCE object representing the processed sublibrary}
}
\value{
SCE output and reports from the split-seq run
}
\description{
generate the sce downsampling statistics
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
