% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{split_stats_output}
\alias{split_stats_output}
\title{intake raw SCE and report global and sample level stats}
\usage{
split_stats_output(sce_split, output_folder, exp_name)
}
\value{
a SCE with stats baked into metadata
}
\description{
intake raw SCE and report global and sample level stats
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
