% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{write_sce_split_lab_filt_stats}
\alias{write_sce_split_lab_filt_stats}
\title{intake complete SCE and write out to sce (.rds) and .H5ad}
\usage{
write_sce_split_lab_filt_stats(sce_split, output_folder, exp_name)
}
\arguments{
\item{sce_split}{a SCE object representing the processed sublibrary}
}
\value{
null - writes sce data to file
}
\description{
intake complete SCE and write out to sce (.rds) and .H5ad
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
