% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{raw_bc_hist}
\alias{raw_bc_hist}
\title{generate hist from raw bc_data}
\usage{
raw_bc_hist(sub_lib_fp, output_folder, exp_name)
}
\value{
complete exit value
}
\description{
generate hist from raw bc_data
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
