% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_split_pipe.R
\name{run_split_pipe}
\alias{run_split_pipe}
\title{Run the split-seq pipeline}
\usage{
run_split_pipe(
  mode = "single",
  n_sublibs = 1,
  data_folder,
  output_folder,
  filtering_mode = "knee",
  filter_value = 1000,
  fastq_path,
  rt_bc = "../test_data_sp_5_miseq/barcodes_v1.csv",
  lig_bc = "../test_data_sp_5_miseq/barcodes_v1.csv",
  sample_map = "../test_data_sp_5_miseq/cell_metadata.xlsx"
)
}
\arguments{
\item{mode}{a string of either 'single' or 'merge'}
}
\value{
SCE outputs, summary figures and reports from zUMIs processing of split-seq FASTQ data
}
\description{
Run the split-seq pipeline
}
\author{
James Opzoomer \email{james.opzoomer@gmail.com}
}
